Move to scripts, use script tools, wrap in try/except
