+++ kpf/linking_table.yml
@@ -127,8 +127,6 @@ links:
     SetFVCExpTime:
         cmd: fvc.SetFVCExpTime.SetFVCExpTime
     # Guider
-    CollectGuiderDarkCubes:
-        cmd: guider.CollectGuiderDarkCubes.CollectGuiderDarkCubes
     DisplayGuiderContinuous:
         cmd: guider.DisplayGuiderContinuous.DisplayGuiderContinuous
     GetGaiaStars:
@@ -166,6 +164,8 @@ links:
         cmd: scripts.CleanupAfterCalibrations.CleanupAfterCalibrations
     CleanupAfterScience:
         cmd: scripts.CleanupAfterScience.CleanupAfterScience
+    CollectGuiderDarkCubes:
+        cmd: scripts.CollectGuiderDarkCubes.CollectGuiderDarkCubes
     ConfigureForAcquisition:
         cmd: scripts.ConfigureForAcquisition.ConfigureForAcquisition
     ConfigureForCalibrations:
